:fire: Remove complex source handling
