+++ commander/src/commands/message/encrypt.ts
@@ -19,31 +19,28 @@ import { flags as flagParser } from '@oclif/command';
 import BaseCommand from '../../base';
 import { ValidationError } from '../../utils/error';
 import { flags as commonFlags } from '../../utils/flags';
-import { getInputsFromSources, InputFromSourceOutput } from '../../utils/input';
+import {
+	getPassphraseFromPrompt,
+	isFileSource,
+	readFileSource,
+} from '../../utils/reader';
 
 interface Args {
 	readonly message?: string;
 	readonly recipientPublicKey: string;
 }
 
-const processInputs = (recipientPublicKey: string, message?: string) => ({
-	passphrase,
-	data,
-}: InputFromSourceOutput) => {
-	const targetMessage = message || data;
-	if (!targetMessage) {
+const processInputs = (
+	recipientPublicKey: string,
+	passphrase: string,
+	message?: string,
+) => {
+	if (!message) {
 		throw new ValidationError('No message was provided.');
-	}
-	if (!passphrase) {
-		throw new ValidationError('No passphrase was provided.');
 	}
 
 	return {
-		...encryptMessageWithPassphrase(
-			targetMessage,
-			passphrase,
-			recipientPublicKey,
-		),
+		...encryptMessageWithPassphrase(message, passphrase, recipientPublicKey),
 		recipientPublicKey,
 	};
 };
@@ -86,19 +83,18 @@ export default class EncryptCommand extends BaseCommand {
 		if (!message && !messageSource) {
 			throw new ValidationError('No message was provided.');
 		}
+		const passphrase =
+			passphraseSource ?? (await getPassphraseFromPrompt('passphrase', true));
+		const dataFromSource =
+			messageSource && isFileSource(messageSource)
+				? await readFileSource(messageSource)
+				: messageSource;
 
-		const inputs = await getInputsFromSources({
-			passphrase: {
-				source: passphraseSource,
-				repeatPrompt: true,
-			},
-			data: message
-				? undefined
-				: {
-						source: messageSource,
-				  },
-		});
-		const result = processInputs(recipientPublicKey, message)(inputs);
+		const result = processInputs(
+			recipientPublicKey,
+			passphrase,
+			message || dataFromSource,
+		);
 		this.print(result);
 	}
 }
